Move tokens down and fix up require paths
This doesn't actually rename the classes yet, but moves everything into
the files where they will be. The next commit will add agent.rb, etc.,
to finish the transition.
