+++ lib/checkpoint/agent/token.rb
