+++ lib/checkpoint/credential/token.rb
