+++ lib/checkpoint/resource/token.rb
