+++ spec/checkpoint/permits_spec.rb
@@ -2,9 +2,9 @@
 
 require 'sequel_helper'
 
-require 'checkpoint/agent'
-require 'checkpoint/credential'
-require 'checkpoint/resource'
+require 'checkpoint/agent/token'
+require 'checkpoint/credential/token'
+require 'checkpoint/resource/token'
 require 'checkpoint/permits'
 
 RSpec.describe Checkpoint::Permits, DB: true do
